ia3

+++ andre/projects/ia/ia3/Environment.java
@@ -1,16 +1,12 @@
-
-// This class provides a stubbed-out environment.
-// You are expected to implement the methods.
-// Accessing an undefined variable should throw an exception.
-
+/**
+ * A referencing environment for bindings
+ */
 public class Environment {
 
-	
-    public int put(String var, int val) { 
+    public double put(String var, int val) {
       	return -1;
     }
-    public int get(int pos, String var) throws EvalException { 
+    public double get(int pos, String var) throws EvalException {
     		return -1;
     }
-    
 }
